Moved SerdeJSON Error to RedisError definition to types.rs
Moved json macro declaration to json.rs

Renamed RedisJson and RedisModule to Json and Module respectfully
